Factor out commitTsGetter interface
