+++ lock-api/src/main/java/com/palantir/lock/client/RemoteTimelockServiceAdapter.java
@@ -37,7 +37,7 @@ public final class RemoteTimelockServiceAdapter implements TimelockService, Auto
     private final NamespacedConjureTimelockService conjureTimelockService;
     private final LockLeaseService lockLeaseService;
     private final TransactionStarter transactionStarter;
-    private final CommitTimestampGetter commitTimestampGetter;
+    private final BatchingCommitTimestampGetter batchingCommitTimestampGetter;
 
     private RemoteTimelockServiceAdapter(
             NamespacedTimelockRpcClient rpcClient,
@@ -47,7 +47,8 @@ public final class RemoteTimelockServiceAdapter implements TimelockService, Auto
         this.rpcClient = rpcClient;
         this.lockLeaseService = LockLeaseService.create(conjureTimelockService, leaderTimeGetter);
         this.transactionStarter = TransactionStarter.create(lockLeaseService, lockWatchEventCache);
-        this.commitTimestampGetter = CommitTimestampGetter.create(lockLeaseService, lockWatchEventCache);
+        this.batchingCommitTimestampGetter =
+                BatchingCommitTimestampGetter.create(lockLeaseService, lockWatchEventCache);
         this.conjureTimelockService = conjureTimelockService;
     }
 
@@ -73,7 +74,7 @@ public final class RemoteTimelockServiceAdapter implements TimelockService, Auto
 
     @Override
     public long getCommitTimestamp(long startTs, LockToken commitLocksToken) {
-        return commitTimestampGetter.getCommitTimestamp(startTs, commitLocksToken);
+        return batchingCommitTimestampGetter.getCommitTimestamp(startTs, commitLocksToken);
     }
 
     @Override
@@ -131,6 +132,6 @@ public final class RemoteTimelockServiceAdapter implements TimelockService, Auto
     @Override
     public void close() {
         transactionStarter.close();
-        commitTimestampGetter.close();
+        batchingCommitTimestampGetter.close();
     }
 }
